refactor(contributors): dropdown content moved to separate file
